requirements.txt
requests
Pillow
gradio_button_lm
pixverse-ai-free==1.6.0
beautifulsoup4
oss2
requests
opencv-python